requirements.txt
fiftyone
numpy
pillow
ipython
pycocotools
matplotlib